JBIDE-29047: Extract the generic functionality from the experimental runtime plugin   - Use class 'org.jboss.tools.hibernate.orm.runtime.common.IFacade' instead of 'org.jboss.tools.hibernate.runtime.common.IFacade' in class 'org.jboss.tools.hibernate.orm.runtime.exp.internal.util.GenericFacadeFactory'   - Use cclass 'org.jboss.tools.hibernate.orm.runtime.common.IFacade' instead of 'org.jboss.tools.hibernate.runtime.common.IFacade' in the test classes of test plugin 'org.jboss.tools.hibernate.orm.runtime.exp.test'
